+++ docs/CURRENT-STATE.md
@@ -2,6 +2,12 @@
 
 ## Infrastructure marker visibility
 
+Datacenter/dam registration uses fresh reusable factories with the application's
+existing context, overlay and render functions. The scoped package exports do
+not import standalone application globals; see
+[the infrastructure interface](INFRASTRUCTURE-LAYERS.md) for lifecycle and asset
+requirements.
+
 Local GeoJSON layers coalesce concurrent enables into one load. Disabling while
 loading keeps the result hidden; destruction aborts the fetch and rejects late
 parse/add results. Destruction and failed setup remove owned context records,
