+++ src/data/localLayers.js
@@ -1,38 +1,9 @@
-import { createLocalGeoJsonLayer } from './localGeojson.js';
+import { localGeoJsonServices } from './localGeojson.js';
+import { createInfrastructureLayers } from './infrastructure.js';
 import { createFirmsHeatmapLayer } from './firmsHeatmap.js';
 import submarineCablesLayer from './telegeographySubmarineCables.js';
 
-// Use Vite's ?url import to properly resolve these assets in dev and build
-import datacentersUrl from './local_data/datacenters/datacenters.geojsonl?url';
-import damsUrl from './local_data/dams/dams.geojsonl?url';
-
-/**
- * Registry of local GeoJSON datasets.
- * These are lazily loaded natively into Cesium when enabled.
- */
-const datacenters = createLocalGeoJsonLayer({
-  id: 'local-datacenters',
-  url: datacentersUrl,
-  name: 'Datacenters',
-  color: '#00ffff', // Cyan
-  icon: '▣',
-  source: 'Local',
-  labels: true,
-  labelMax: 700,
-  labelGridPx: 138,
-});
-
-const dams = createLocalGeoJsonLayer({
-  id: 'local-dams',
-  url: damsUrl,
-  name: 'Dams',
-  color: '#0088ff', // Blue
-  icon: '▰',
-  source: 'USACE',
-  labels: true,
-  labelMax: 900,
-  labelGridPx: 132,
-});
+const [datacenters, dams] = createInfrastructureLayers(localGeoJsonServices);
 
 // Live NASA FIRMS fires (VIIRS ×3 NRT via the /api/firms proxy). The id keeps
 // the historical `local-` prefix for persistence + voice-tool-enum compat,
